api.model rename after move
