+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/assocends/AllAssociationEndTests.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.api.model.tests.assocends;
+package hu.elte.txtuml.api.model.assocends;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
