+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/error/deletion/AllDeletionErrorTests.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.api.model.tests.error.deletion;
+package hu.elte.txtuml.api.model.error.deletion;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
