+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/error/deletion/CannotBeDeletedTest.java
@@ -1,9 +1,9 @@
-package hu.elte.txtuml.api.model.tests.error.deletion;
+package hu.elte.txtuml.api.model.error.deletion;
 
 import hu.elte.txtuml.api.model.Action;
-import hu.elte.txtuml.api.model.tests.base.SimpleModelTestsBase;
-import hu.elte.txtuml.api.model.tests.models.SimpleModel.A_B;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.SimpleModelTestsBase;
+import hu.elte.txtuml.api.model.models.SimpleModel.A_B;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
