+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/error/deletion/StartingDeletedTest.java
@@ -1,8 +1,8 @@
-package hu.elte.txtuml.api.model.tests.error.deletion;
+package hu.elte.txtuml.api.model.error.deletion;
 
 import hu.elte.txtuml.api.model.Action;
-import hu.elte.txtuml.api.model.tests.base.SimpleModelTestsBase;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.SimpleModelTestsBase;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
