+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/error/other/ChangingLockedExecutionTimeMultiplierTest.java
@@ -1,11 +1,11 @@
-package hu.elte.txtuml.api.model.tests.error.other;
+package hu.elte.txtuml.api.model.error.other;
 
 import hu.elte.txtuml.api.model.Action;
 import hu.elte.txtuml.api.model.ModelExecutor;
-import hu.elte.txtuml.api.model.tests.base.SimpleModelTestsBase;
-import hu.elte.txtuml.api.model.tests.models.SimpleModel.A_B;
-import hu.elte.txtuml.api.model.tests.models.SimpleModel.Sig;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.SimpleModelTestsBase;
+import hu.elte.txtuml.api.model.models.SimpleModel.A_B;
+import hu.elte.txtuml.api.model.models.SimpleModel.Sig;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 import hu.elte.txtuml.api.stdlib.Timer;
 
 import org.junit.Test;
