+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/models/AssociationsModel.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.api.model.tests.models;
+package hu.elte.txtuml.api.model.models;
 
 import hu.elte.txtuml.api.model.Association;
 import hu.elte.txtuml.api.model.Model;
