+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/shutdown/AllShutdownTests.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.api.model.tests.shutdown;
+package hu.elte.txtuml.api.model.shutdown;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
