+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/shutdown/ShutdownQueueTest.java
@@ -1,9 +1,9 @@
-package hu.elte.txtuml.api.model.tests.shutdown;
+package hu.elte.txtuml.api.model.shutdown;
 
 import hu.elte.txtuml.api.model.ModelExecutor;
-import hu.elte.txtuml.api.model.tests.base.SimpleModelTestsBase;
-import hu.elte.txtuml.api.model.tests.util.MutableBoolean;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.SimpleModelTestsBase;
+import hu.elte.txtuml.api.model.util.MutableBoolean;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 
 import org.junit.Assert;
 import org.junit.Test;
