+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/statemachine/ChoiceTest.java
@@ -1,9 +1,9 @@
-package hu.elte.txtuml.api.model.tests.statemachine;
+package hu.elte.txtuml.api.model.statemachine;
 
 import hu.elte.txtuml.api.model.Action;
-import hu.elte.txtuml.api.model.tests.base.ChoiceModelTestsBase;
-import hu.elte.txtuml.api.model.tests.models.ChoiceModel.Sig;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.ChoiceModelTestsBase;
+import hu.elte.txtuml.api.model.models.ChoiceModel.Sig;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
