+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/statemachine/EntryExitEffectTest.java
@@ -1,10 +1,10 @@
-package hu.elte.txtuml.api.model.tests.statemachine;
+package hu.elte.txtuml.api.model.statemachine;
 
 import hu.elte.txtuml.api.model.Action;
-import hu.elte.txtuml.api.model.tests.base.TransitionsModelTestsBase;
-import hu.elte.txtuml.api.model.tests.models.TransitionsModel.Sig1;
-import hu.elte.txtuml.api.model.tests.models.TransitionsModel.Sig2;
-import hu.elte.txtuml.api.model.tests.util.SeparateClassloaderTestRunner;
+import hu.elte.txtuml.api.model.base.TransitionsModelTestsBase;
+import hu.elte.txtuml.api.model.models.TransitionsModel.Sig1;
+import hu.elte.txtuml.api.model.models.TransitionsModel.Sig2;
+import hu.elte.txtuml.api.model.util.SeparateClassloaderTestRunner;
 
 import org.junit.Assert;
 import org.junit.Test;
