+++ tests/hu.elte.txtuml.api.model.tests/src/hu/elte/txtuml/api/model/util/SeparateClassloaderTestRunner.java
@@ -1,4 +1,4 @@
-package hu.elte.txtuml.api.model.tests.util;
+package hu.elte.txtuml.api.model.util;
 
 import java.net.URLClassLoader;
